Do not use title to determine fingerprint
As it was, if an error contained a different string, even though it was
from the same source, the fingerprint would be different and the errors
would not be grouped in Rollbar. Now we guarantee that the hash is
always the same for the same filename, method name, and line number.